Added drush wrapper command.
This drush wrapper allows us to run drush on servers where we're Drupal's
settings are being auto-set based on the environment variables for the site's
virtual host config. Since drush runs from the command line, we need to
manually set the environment variables for APP_ENV and APP_NAME each time
drush runs or any command requiring access to the database will fail as it
won't know what the db name or db user is.