Create a new file instead of using stdout.
I think most of the time the behaviour we want would be to replace the
input file with the same file except without the definer clauses. I don't
want to do that because I don't trust the script to not mess up the file
and these files could be important backups. To get closer to that, though,
it now creates a file with the same name as the old one but with
-definer-fixed appended to the base name so that we don't lose the
original backup.

If the file ends in .gz, it assumes it is gzipped and ungzips the input as
it reads it and outputs a gzipped file.