Adjusted for new FixDefiner and PHP 5.3.
It looks like the bindings are different for anonymous functions between
PHP 5.3 and 5.5, so you can't reference $this in an anonymous function
created inside a method in PHP 5.3. So I'm passing it in now which works
for both.

Also, needed to adjust the RestoreLatestS3Backup for the new FixDefiner
which can handle gzip.