Fixed problem with chunks crossing line endings.
If a chunk read across an EOL then that could cause the regex to not match
when it should because not all of the line is there to compare. I changed
it to use fgets which always just reads to EOL (or EOF). This is probably
not as fast, but it should work all the time. The chunk read was failing
intermittently causing DEFINERs to be left in the file.