Made the restore backup work better with B&M files.
The restore backup stuff was assuming that the directory inside the tarballs
created by backup and migrate was alwas named with the same name as the tarball
itself. This is not true if there is a '.' in the name of the tarball before
the timestamp. I rewrote it to just look for a directory called 'files' in the
tarball and just use that if it finds it. This should keep it working in more
cases, I think.