Make site root detection better. Issue #21143.
Previously, we were handling root path detection in multiple different ways.
Also, we were doing it for every command, regardless of whether the command
needed it or not, which is just a waste. A similar thing was happening for
environment detection (wordpress or drupal).

This commit standardizes the handling of this information in a new class that
can be accessed from any command that needs this information.