Added endpoint parameter.
Education Outside doesn't use the standard S3 region, so we need to
specify an endpoint.

I also moved some of the stuff that was in the constructor into the run
command because the constructor gets run when requesting help for the
command, but sometimes there's an error in the constructor in which case
you can't see the help even though you don't actually need to run the
stuff in the constructor.